requirements.txt
django~=3.2.9
pillow
psycopg2-binary
djoser
django-haystack~=3.1.1
djangorestframework~=3.13.1
web3~=5.30.0
web3auth~=1.4.0
requests~=2.28.1